xilpm: versal: server: Fix MISRA 10.8 violation
Fix below MISRA-C violation

misra_c_2012_rule_10_8_violation:
The value of composite expression shall not be cast to a different
essential type category or a wider essential type.


Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyotheeswar.reddy.mutthareddyvari@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_common.c
@@ -75,7 +75,7 @@ void XPm_DumpMemUsage(void)
 {
 	xil_printf("Total buffer size = %d bytes\n\r", MAX_BYTEBUFFER_SIZE);
 	xil_printf("Used = %d bytes\n\r", FreeBytes - ByteBuffer);
-	xil_printf("Free = %d bytes\n\r", MAX_BYTEBUFFER_SIZE - (u32)(FreeBytes - ByteBuffer));
+	xil_printf("Free = %d bytes\n\r", MAX_BYTEBUFFER_SIZE - (u32)FreeBytes - (u32)ByteBuffer);
 	xil_printf("\n\r");
 }
